Front moved to the separate actor.

+++ new-core/src/main/java/com/spbsu/datastream/core/range/AddressedMessage.java
@@ -3,19 +3,19 @@ package com.spbsu.datastream.core.range;
 public final class AddressedMessage<T> {
   private final int hash;
 
-  private final int tick;
+  private final long tick;
 
   private final T payload;
 
   public AddressedMessage(final T payload,
                           final int hash,
-                          final int tick) {
+                          final long tick) {
     this.payload = payload;
     this.hash = hash;
     this.tick = tick;
   }
 
-  public int tick() {
+  public long tick() {
     return this.tick;
   }
